Move providers to inner classes per check and update ModularFramework
