+++ build.gradle
@@ -30,7 +30,7 @@ dependencies {
     compile 'org.apache.commons:commons-math3:3.6.1'
     compile 'org.spongepowered:spongeapi:6.0.0-SNAPSHOT'
     compile 'com.zaxxer:HikariCP:2.6.0'
-    compile 'com.me4502:ModularFramework:1.8'
+    compile 'com.me4502:ModularFramework:1.8.3'
 }
 
 if (JavaVersion.current().isJava8Compatible()) {
